requirements.txt
emoji>=2.11.0
pillow>=10.3.0
requests>=2.31.0
Unidecode>=1.3.8
moviepy>=1.0.3
